fix(core): clean up trivial clippy lints under --all-targets --all-features (#180)
PR 1 of 5 in the #180 series to close the gap between CI's clippy gate
and `cargo clippy --all-targets --all-features -- -D warnings`. Local
baseline drops from 60 errors to 43; remaining lints partition cleanly
into the PR 2-5 backlog (FFI shims, field-reassign-with-default, SIMD
too_many_arguments, CI tightening).

Trivial mechanical fixes only — zero behavior change:
- ar/image_proc.rs: silence unused-vars in luma_hist_and_box_filter_with_bias
  (variables used only under specific target_feature gates); convert
  manual range check to (100..200).contains(&thresh).
- ar/math.rs, ar/pattern.rs: move items above the test module so
  clippy::items_after_test_module is satisfied.
- ar/matrix.rs, ar/pattern.rs, ar2/feature_map.rs, ar2/image_set.rs:
  convert index-only loops to iterator / enumerate forms
  (clippy::needless_range_loop).
- ar/bch.rs: drop no-op `as i32` cast.
- arlog.rs: drop needless `fn main` in doctest.
- version.rs: use char-array split pattern.
- kpm/freak/homography.rs: drop needless `*pt` deref.
- kpm/freak/math.rs: convert manual inclusive range to (0.0..=360.0).contains.

Adds .claude/issue-180-plan.md with the brainstormed 5-PR plan,
Decision Log, assumptions, and risks for the remaining series.

+++ crates/core/src/ar/bch.rs
@@ -484,7 +484,7 @@ pub(crate) mod test_helpers {
             // elements stored as i32 (0 represents zero element).
             let mut m_coeffs: Vec<i32> = vec![1]; // M(x) = 1 in GF(2^7)
             for &i in &coset {
-                let alpha_i = BCH_127_ALPHA_TO[i] as i32;
+                let alpha_i = BCH_127_ALPHA_TO[i];
                 // Multiply m_coeffs by (x + α^i) (subtraction == addition in GF(2)).
                 let mut new_m = vec![0i32; m_coeffs.len() + 1];
                 for (idx, &c) in m_coeffs.iter().enumerate() {
